+++ ts/core/index.ts
@@ -1,4 +1,3 @@
-import 'bootstrap/dist/js/bootstrap.bundle.min';
 import './lang/string';
 import './lang/number';
 import './lang/date';
@@ -8,5 +7,3 @@ import * as _Rosie from './mixins';
 export const Rosie = { ..._Rosie };
 
 export * from './types';
-export * from './ajax';
-export * from './data';